requirements.txt
streamlit
mysql-connector-python
python-dotenv
requests
pandas
